requirements.txt
ultralytics==8.0.196
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
gunicorn==21.2.0
setuptools==69.0.0
